Move RegionSelector out of mplcpp
It's not matplotlib, so should probably exist in a module of its own.

RE #33735
