+++ qt/widgets/CMakeLists.txt
@@ -5,3 +5,4 @@ add_subdirectory(plotting)
 add_subdirectory(instrumentview)
 add_subdirectory(plugins)
 add_subdirectory(testhelpers)
+add_subdirectory(regionselector)
